New Features and Streamlining
Reduced the CSS for the mobile menu button considerably and removed the
:before styles. Added a new 'animation' and 'padding' feature. Beta:
Added resize() to adjust the top-padding on the body if browser is
resized for mobile. Also made it work properly for the initial browser
dimensions and made it reset to the default padding when resized to
something other than mobile dimensions. Current breakpoint is 480px.